Cnc: added a mechanism to wait for responses
The interface supports both synchronous, asynchronous and
fire&forget type of messages. The operations themselves are
first-class citizens and after construction they shall be
passed to Cnc.wait, .async or .ignore, to be actually sent
down the wire.

There are no combinator functions available for the operations
and they probably won't be added before a use case for them
appears.