Scan: use another approach with linear regression
Instead of considering all the offsets in an image as parameters to
the linear regression, consider each (offset, z-offset) of each
different angle as separate sets of teaching material. This allows to
handle missing samples and also allows in future to consider the
trustworthiness of certain offsets (for example, of the
laser-projected line goes towards the laser, it is not a good
indicator).